requirements.txt
Flask
mysql
azure-mgmt-resource
azure-mgmt-rdbms
azure-identity
mysql-connector-python
adal
msrestazure
pymysql